First HTML and CSS layout
First layout of the Contact++ app.  I added a basic HTML

structure with a table outlining the different fields of our future data set.

I included already three buttons:
  - Edit
  - Delete
  - Add contact

I started to style the elements to get an idea of how it will feel. This will most likely to be changed. :)